requirements.txt
asgiref==3.7.2
bleach==6.0.0
Django==4.2.1
django-appconf==1.0.5
django-compressor==4.4
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==5.0
django-simple-captcha==0.5.18
django-summernote==0.8.20.0
django-taggit==4.0.0
Pillow==10.0.0
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
webencodings==0.5.1
